move utils, delete shared
