+++ next.config.mjs
@@ -20,23 +20,6 @@ const nextConfig = {
       },
     ],
   },
-  modularizeImports: {
-    "@/components/?(((\\w*)?/?)*)": {
-      transform: "@/components/{{ matches.[1] }}/{{member}}",
-      preventFullImport: true,
-      skipDefaultConversion: true,
-    },
-    "@utils/?(((\\w*)?/?)*)": {
-      transform: "@utils/{{ matches.[1] }}/{{member}}",
-      preventFullImport: true,
-      skipDefaultConversion: true,
-    },
-    "@/features/?(((\\w*)?/?)*)": {
-      transform: "@/features/{{ matches.[1] }}/{{member}}",
-      preventFullImport: true,
-      skipDefaultConversion: true,
-    },
-  },
   transpilePackages: ["monaco-editor"],
   /**
    * @param {import("webpack").Configuration} config
@@ -58,6 +41,15 @@ const nextConfig = {
   },
   experimental: {
     ppr: true,
+    optimizePackageImports: [
+      "@/features/auth",
+      "@/features/about",
+      "@/features/rankings",
+      "@/features/problems",
+      "@/features/submissions",
+      "@/utils",
+      "@/components",
+    ],
   },
 };
