+++ src/components/Input/index.tsx
@@ -1,8 +1,8 @@
-import { clsx, createNamespaceComponent } from "@utils/shared";
 import type { ComponentPropsWithoutRef } from "react";
 import { forwardRef } from "react";
 
 import type { Optional, RequireFields } from "@/types";
+import { clsx, createNamespaceComponent } from "@/utils";
 
 import { InputErrorText } from "./InputErrorText";
 import type { LabelProps } from "./InputLabel";
