+++ src/components/Score.tsx
@@ -1,7 +1,7 @@
-import { clsx, getVerdictTextClass } from "@utils/shared";
 import Link from "next/link";
 
 import type { DetailedScoreInfo } from "@/types";
+import { clsx, getVerdictTextClass } from "@/utils";
 
 interface ComponentExtraProps {
   className?: string;
