+++ src/features/problems/actions/actionCreateProblem.ts
@@ -2,10 +2,24 @@
 
 import prisma from "@database/prisma/instance";
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
-import { createProblemSchema } from "@schemas/createProblemSchema";
 import { cookies } from "next/headers";
+import { z } from "zod";
 
 import { getUser } from "@/features/auth";
+
+export const createProblemSchema = z.object({
+  code: z.string().max(20, "Code must contain at most 20 letters").optional(),
+  name: z.string().max(255, "Name must contain at most 255 letters").optional(),
+  rid: z.number(),
+});
+
+export type CreateProblemSchemaType = z.infer<typeof createProblemSchema>;
+
+export interface CreateProblemResponse {
+  status: "OK" | "DUPLICATED" | "FAILED";
+  message: string;
+  code: string;
+}
 
 export interface CreateProblemFormState {
   code_messages: string[];
