+++ src/features/problems/queries/getProblem.ts
@@ -1,9 +1,9 @@
 import prisma from "@database/prisma/instance";
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library";
-import { parseJudge } from "@utils/shared";
 import { unstable_cache } from "next/cache";
 
 import type { DetailedProblemInfo } from "@/features/problems";
+import { parseJudge } from "@/utils";
 
 export const getProblem = async (code: string) => {
   try {
