+++ src/utils/createNamespaceComponent.ts
