+++ src/utils/getResult.ts
