+++ src/utils/getVerdictTextClass.ts
