+++ tsconfig.json
@@ -15,42 +15,8 @@
     "jsx": "preserve",
     "incremental": true,
     "paths": {
-      "@schemas/*": ["./src/shared/schemas/*"],
-      "@utils/*": ["./src/shared/utils/*"],
       "@database/*": ["./src/database/*"],
       "@/*": ["./src/*"],
-      "@/features/about/*":[
-        "./src/features/about/types/*",
-        "./src/features/about/assets/*"
-      ],
-      "@/features/auth/*":[
-        "./src/features/auth/api/*",
-        "./src/features/auth/components/*",
-        "./src/features/auth/queries/*",
-        "./src/features/auth/actions/*",
-        "./src/features/auth/types",
-      ],
-      "@/features/problems/*":[
-        "./src/features/problems/api/*",
-        "./src/features/problems/components/*",
-        "./src/features/problems/queries/*",
-        "./src/features/problems/actions/*",
-        "./src/features/problems/types",
-      ],
-      "@/features/submissions/*":[
-        "./src/features/submissions/api/*",
-        "./src/features/submissions/components/*",
-        "./src/features/submissions/queries/*",
-        "./src/features/submissions/actions/*",
-        "./src/features/submissions/types",
-      ],
-      "@/features/rankings/*":[
-        "./src/features/rankings/api/*",
-        "./src/features/rankings/components/*",
-        "./src/features/rankings/queries/*",
-        "./src/features/rankings/actions/*",
-        "./src/features/rankings/types",
-      ],
     },
     "types": ["react/experimental"],
     "plugins": [
@@ -59,6 +25,6 @@
       }
     ]
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/shared/utils/shared/md5.js"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "src/utils/md5.js"],
   "exclude": ["node_modules"]
 }
